test(e2e): clean up search e2e structure + verify reflow repro (#4645)
Infra/structure:
- recursive spec glob (specs/**/*.spec.ts) so specs can live in subfolders
- move search spec to e2e/specs/search/ + add e2e/specs/search/README.md
- extract e2e/helpers/search.ts (openSearch / fixture gate / close helper)
- test:e2e:search-bugs sets SCREENPIPE_PORT so the harness app boots its own
  server alongside a dev machine's screenpipe on :3030 (uses the documented
  escape hatch instead of colliding)

Verified locally (node 22, webkit, isolated port + seeded DB): the reflow test
fails for the right reason — grid stays at 3 columns on a 1680px window
(hard-coded grid-cols-3), Expected > 3 / Received 3.

Scope trimmed to what actually reproduces: dropped the 60vh and highlight
assertions — 60vh passed (not a real repro) and the highlight over-match did
not reproduce via the rendered overlay. Those + the timeline-side bugs are
documented in the spec/README for follow-up rather than shipped as
passing/empty 'repro' tests.

Refs #4645
